refactor preprocessors
